nix: allow port forwarding configuration
We add a nixos module to the testing modules. This allows configuring
ports for port forwarding for the dashboard and the serialproxy.

This way, users can access the dashboard and the web serial console on
the host when running the OpenStack controller in a VM.

On-behalf-of: SAP stefan.kober@sap.com